refactor: Move dockerflow router to merino-common
