+++ merino-common/merino_common/routers/__init__.py
@@ -1,0 +1,1 @@
+"""Shared FastAPI routers for Merino services."""
